refactor app folders
