+++ lib/main.dart
@@ -1,5 +1,8 @@
-import 'package:daily_routine_app_isar/screens/main_screen.dart';
 import 'package:flutter/material.dart';
+
+import 'src/screens/main_screen.dart';
+import 'src/utils/dimens.dart';
+import 'src/utils/themes.dart';
 
 void main() {
   runApp(const MyApp());
@@ -15,9 +18,42 @@ class MyApp extends StatelessWidget {
         debugShowCheckedModeBanner: false,
         title: 'Routine app',
         theme: ThemeData(
-          colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
-          useMaterial3: true,
-        ),
+            colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
+            useMaterial3: true,
+            inputDecorationTheme: InputDecorationTheme(
+              border: const OutlineInputBorder(
+                  borderSide: BorderSide(color: Colors.grey),
+                  borderRadius:
+                      BorderRadius.all(Radius.circular(AppDimens.medium))),
+              enabledBorder: const OutlineInputBorder(
+                  borderSide: BorderSide(color: Colors.grey),
+                  borderRadius:
+                      BorderRadius.all(Radius.circular(AppDimens.medium))),
+              focusedBorder: OutlineInputBorder(
+                borderSide: BorderSide(color: AppColors.secondaryColor),
+                borderRadius:
+                    const BorderRadius.all(Radius.circular(AppDimens.medium)),
+              ),
+            ),
+            elevatedButtonTheme: const ElevatedButtonThemeData(
+              style: ButtonStyle(
+                shape: MaterialStatePropertyAll(RoundedRectangleBorder(
+                    borderRadius:
+                        BorderRadius.all(Radius.circular(AppDimens.medium)))),
+              ),
+            ),
+            dropdownMenuTheme: DropdownMenuThemeData(
+                inputDecorationTheme: InputDecorationTheme(
+              border: const OutlineInputBorder(
+                  borderSide: BorderSide(color: Colors.grey),
+                  borderRadius:
+                      BorderRadius.all(Radius.circular(AppDimens.medium))),
+              focusedBorder: OutlineInputBorder(
+                borderSide: BorderSide(color: AppColors.secondaryColor),
+                borderRadius:
+                    const BorderRadius.all(Radius.circular(AppDimens.medium)),
+              ),
+            ))),
         home: const MainScreen());
   }
 }
