+++ lib/src/utils/dimens.dart
@@ -1,0 +1,6 @@
+class AppDimens {
+  AppDimens._();
+  static const double small = 8;
+  static const double medium = 16;
+  static const double large = 24;
+}
