+++ lib/src/utils/themes.dart
@@ -1,0 +1,8 @@
+import 'package:flutter/material.dart';
+
+class AppColors {
+  AppColors._();
+
+  static const Color primaryColor = Colors.indigo;
+  static Color secondaryColor = const Color.fromARGB(128, 63, 81, 181);
+}
